+++ tests/spcs/test_compute_pool.py
@@ -53,7 +53,7 @@ def test_create(mock_execute_query):
     comment = "'test comment'"
     cursor = Mock(spec=SnowflakeCursor)
     mock_execute_query.return_value = cursor
-    result = ComputePoolManager().create_from_params(
+    result = ComputePoolManager().create(
         pool_name=pool_name,
         min_nodes=min_nodes,
         max_nodes=max_nodes,
@@ -82,9 +82,7 @@ def test_create(mock_execute_query):
     assert result == cursor
 
 
-@patch(
-    "snowflake.cli._plugins.spcs.compute_pool.manager.ComputePoolManager.create_from_params"
-)
+@patch("snowflake.cli._plugins.spcs.compute_pool.manager.ComputePoolManager.create")
 def test_create_pool_cli_defaults(mock_create, runner):
     result = runner.invoke(
         [
@@ -99,14 +97,19 @@ def test_create_pool_cli_defaults(mock_create, runner):
     assert result.exit_code == 0, result.output
     mock_create.assert_called_once_with(
         pool_name="test_pool",
+        min_nodes=1,
+        max_nodes=1,
         instance_family="test_family",
+        auto_resume=True,
+        initially_suspended=False,
+        auto_suspend_secs=3600,
+        comment=None,
         replace=False,
+        if_not_exists=False,
     )
 
 
-@patch(
-    "snowflake.cli._plugins.spcs.compute_pool.manager.ComputePoolManager.create_from_params"
-)
+@patch("snowflake.cli._plugins.spcs.compute_pool.manager.ComputePoolManager.create")
 def test_create_pool_cli(mock_create, runner):
     result = runner.invoke(
         [
@@ -149,7 +152,7 @@ def test_create_pool_cli(mock_create, runner):
 def test_create_compute_pool_already_exists(mock_handle, mock_execute):
     pool_name = "test_pool"
     mock_execute.side_effect = SPCS_OBJECT_EXISTS_ERROR
-    ComputePoolManager().create_from_params(
+    ComputePoolManager().create(
         pool_name=pool_name,
         min_nodes=1,
         max_nodes=1,
@@ -173,7 +176,7 @@ def test_create_compute_pool_already_exists(mock_handle, mock_execute):
 def test_create_compute_pool_if_not_exists(mock_execute_query):
     cursor = Mock(spec=SnowflakeCursor)
     mock_execute_query.return_value = cursor
-    result = ComputePoolManager().create_from_params(
+    result = ComputePoolManager().create(
         pool_name="test_pool",
         min_nodes=1,
         max_nodes=1,
@@ -243,19 +246,14 @@ def test_create_compute_pool_replace(
 
 
 @patch(EXECUTE_QUERY)
-def test_create_from_project_definition(
-    mock_execute_query, runner, project_directory, mock_cursor, os_agnostic_snapshot
-):
-    mock_execute_query.return_value = mock_cursor(
-        rows=[["Compute pool TEST_COMPUTE_POOL successfully created."]],
-        columns=["status"],
-    )
-
+def test_deploy(mock_execute_query, runner, project_directory):
     with project_directory("spcs_compute_pool"):
-        result = runner.invoke(["spcs", "compute-pool", "create"])
+        result = runner.invoke(["spcs", "compute-pool", "deploy"])
 
         assert result.exit_code == 0, result.output
-        assert result.output == os_agnostic_snapshot
+        assert (
+            "Compute pool 'test_compute_pool' successfully deployed." in result.output
+        )
         expected_query = dedent(
             """\
             CREATE COMPUTE POOL test_compute_pool
@@ -271,25 +269,19 @@ def test_create_from_project_definition(
 
 @patch("snowflake.cli._plugins.object.manager.ObjectManager.execute_query")
 @patch(EXECUTE_QUERY)
-def test_create_from_project_definition_replace(
-    mock_execute_query,
-    mock_execute_query_object_manager,
-    runner,
-    project_directory,
-    mock_cursor,
-    os_agnostic_snapshot,
+def test_deploy_replace(
+    mock_execute_query, mock_execute_query_object_manager, runner, project_directory
 ):
     compute_pool_name = "test_compute_pool"
-    mock_execute_query.return_value = mock_cursor(
-        rows=[["Compute pool TEST_COMPUTE_POOL successfully created."]],
-        columns=["status"],
-    )
 
     with project_directory("spcs_compute_pool"):
-        result = runner.invoke(["spcs", "compute-pool", "create", "--replace"])
+        result = runner.invoke(["spcs", "compute-pool", "deploy", "--replace"])
 
         assert result.exit_code == 0, result.output
-        assert result.output == os_agnostic_snapshot
+        assert (
+            f"Compute pool '{compute_pool_name}' successfully deployed."
+            in result.output
+        )
         expected_query = dedent(
             f"""\
             CREATE COMPUTE POOL {compute_pool_name}
@@ -315,7 +307,7 @@ def test_create_from_project_definition_replace(
 
 
 @patch(EXECUTE_QUERY)
-def test_create_from_project_definition_compute_pool_already_exists(
+def test_deploy_compute_pool_already_exists(
     mock_execute_query, runner, project_directory
 ):
     mock_execute_query.side_effect = ProgrammingError(
@@ -323,7 +315,7 @@ def test_create_from_project_definition_compute_pool_already_exists(
     )
 
     with project_directory("spcs_compute_pool"):
-        result = runner.invoke(["spcs", "compute-pool", "create"])
+        result = runner.invoke(["spcs", "compute-pool", "deploy"])
 
         assert result.exit_code == 1, result.output
         assert (
@@ -332,43 +324,38 @@ def test_create_from_project_definition_compute_pool_already_exists(
         )
 
 
-def test_create_from_project_definition_no_compute_pools(runner, project_directory):
+def test_deploy_no_compute_pools(runner, project_directory):
     with project_directory("empty_project"):
-        result = runner.invoke(["spcs", "compute-pool", "create"])
+        result = runner.invoke(["spcs", "compute-pool", "deploy"])
 
         assert result.exit_code == 1, result.output
         assert "No compute pool project definition found in" in result.output
 
 
-def test_create_from_project_definition_not_existing_entity_id(
-    runner, project_directory
-):
+def test_deploy_not_existing_entity_id(runner, project_directory):
     with project_directory("spcs_compute_pool"):
         result = runner.invoke(
-            ["spcs", "compute-pool", "create", "not_existing_entity_id"]
+            ["spcs", "compute-pool", "deploy", "not-existing-entity-id"]
         )
 
         assert result.exit_code == 2, result.output
         assert (
-            "No 'not_existing_entity_id' entity in project definition file."
+            "No 'not-existing-entity-id' entity in project definition file."
             in result.output
         )
 
 
 @patch(EXECUTE_QUERY)
-def test_create_from_project_definition_multiple_compute_pools_with_entity_id(
-    mock_execute_query, runner, project_directory, mock_cursor, os_agnostic_snapshot
+def test_deploy_multiple_compute_pools_with_entity_id(
+    mock_execute_query, runner, project_directory
 ):
-    mock_execute_query.return_value = mock_cursor(
-        rows=[["Compute pool TEST_COMPUTE_POOL successfully created."]],
-        columns=["status"],
-    )
-
     with project_directory("spcs_multiple_compute_pools"):
-        result = runner.invoke(["spcs", "compute-pool", "create", "test_compute_pool"])
+        result = runner.invoke(["spcs", "compute-pool", "deploy", "test_compute_pool"])
 
         assert result.exit_code == 0, result.output
-        assert result.output == os_agnostic_snapshot
+        assert (
+            "Compute pool 'test_compute_pool' successfully deployed." in result.output
+        )
         expected_query = dedent(
             """\
             CREATE COMPUTE POOL test_compute_pool
@@ -382,11 +369,9 @@ def test_create_from_project_definition_multiple_compute_pools_with_entity_id(
         mock_execute_query.assert_called_once_with(expected_query)
 
 
-def test_create_from_project_definition_multiple_compute_pools(
-    runner, project_directory
-):
+def test_deploy_multiple_compute_pools(runner, project_directory):
     with project_directory("spcs_multiple_compute_pools"):
-        result = runner.invoke(["spcs", "compute-pool", "create"])
+        result = runner.invoke(["spcs", "compute-pool", "deploy"])
 
         assert result.exit_code == 2, result.output
         assert (